Extract tooltip functionality to separate component
